+++ lib/mix/tasks/star_view.install.ex
@@ -141,6 +141,10 @@ if Code.ensure_loaded?(Igniter) do
               import Phoenix.Component, except: [assign: 3]
               import Plug.Conn
 
+              alias MyAppWeb.Components.StarView.Layout
+
+              plug :put_root_layout, html: {Layout, :root}
+
               unquote(verified_routes())
             end
           end